feat: refactor and create simulations

+++ simulations/domain/ipca.py
@@ -1,8 +1,8 @@
 from decimal import Decimal
 from pyparsing import Char
 
-from transactions.domain.helpers.ipca import HelpersIpca
+from simulations.domain.helpers.ipca import HelpersIpca
 
 class TransactionIpca:
-    def __init__(self, invest: Decimal, name: Char, deadline: int) -> None:
+    def __init__(self, invest: int, name: Char, deadline: int) -> None:
         self.helpersIpca = HelpersIpca(invest=invest, name=name, deadline=deadline)
